requirements.txt
playwright==1.40.0
playwright-stealth==1.0.6
google-search-results==2.4.2
google-generativeai==0.3.2
python-dotenv==1.0.0
gradio==4.16.0
huggingface-hub==0.20.0
robotexclusionrulesparser==1.7.1
urllib3==2.1.0
requests==2.31.0
setuptools>=65.0.0
